message

+++ Skills-Overflow-Backend/src/main/java/com/internshipProject/SkillsOverflowBackend/threads/TokenThread.java
@@ -11,9 +11,6 @@ import java.time.LocalDateTime;
 
 public class TokenThread implements Runnable{
 
-//    @Autowired
-//    MailService mailService;
-
     @Autowired
     UserRepository userRepository;
 
@@ -23,18 +20,25 @@ public class TokenThread implements Runnable{
     @Override
     public void run() {
         while(true) {
-            userRepository
-                    .findAll()
-                    .stream()
-                    .filter(user ->
-                            user.getEnabled()
-                                    && LocalDateTime.now().isAfter(user.getVerificationToken().getExpirationDate()))
-                    .forEach(user-> {
-                        tokenRepository.delete(user.getVerificationToken());
-                        userRepository.delete(user);
-                    });
+            if (userRepository.findAll().size()!=0 && tokenRepository.findAll().size()!=0) {
+                userRepository
+                        .findAll()
+                        .forEach(user -> {
+                            if (LocalDateTime.now().isAfter(user.getVerificationToken()
+                                    .getExpirationDate())) {
+                                tokenRepository.delete(user.getVerificationToken());
+                                System.out.println("Token is..." + user.getVerificationToken());
+                                System.out.println("-----------");
+                                if (!user.getEnabled()) {
+                                    userRepository.delete(user);
+                                    System.out.println("User is ..." + user);
+                                }
+                            }
+                        });
+            }
             try {
-                Thread.sleep(1000*3600);
+                Thread.sleep(1000*20);
+                System.out.println("NOT SLEEPING ANYMORE");
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
